chore(deps-dev): Bump psalm/phar from 6.5.1 to 6.14.3

+++ lib/Listener/RegisterTemplateFileCreatorListener.php
@@ -29,6 +29,7 @@ class RegisterTemplateFileCreatorListener implements IEventListener {
 	) {
 	}
 
+	#[\Override]
 	public function handle(Event $event): void {
 		if (!$event instanceof RegisterTemplateCreatorEvent) {
 			return;
@@ -58,7 +59,7 @@ class RegisterTemplateFileCreatorListener implements IEventListener {
 				$odtType->addMimetype('application/vnd.oasis.opendocument.text-template');
 			}
 			$odtType->setIconSvgInline(file_get_contents($appPath . '/img/x-office-document.svg'));
-			$odtType->setRatio(21 / 29.7);
+			$odtType->setRatio((float)21 / 29.7);
 			return $odtType;
 		});
 		$templateManager->registerTemplateFileCreator(function () use ($ooxml, $appPath) {
